1.）dash board cpu 核数耗费改为double类型，原因：cpu请求过低时不易显示.

+++ agent-manager/agent-manager-common/src/main/java/com/didichuxing/datachannel/agentmanager/common/bean/vo/dashboard/DashBoardVO.java
@@ -40,7 +40,7 @@ public class DashBoardVO {
     private Long logCollectPathNum;
 
     @ApiModelProperty(value = "系统全量agent cpu耗费（单位：core）")
-    private Long agentCpuCoresSpend;
+    private Double agentCpuCoresSpend;
 
     @ApiModelProperty(value = "系统全量agent memory耗费（单位：byte）")
     private Long agentMemorySpend;
